Fix long outstanding bug in WITH-TRANSACTION that was arguably making it rather useless.
From the TODO file, augmented:

Within anonymous transactions, circular dependencies are not correctly
serialized.  If an object that is created in the anonymous
transactions links itself to another object, it may fail to restore
correctly.

The issue here is that the anonymous transaction is first completely
read from the transaction log before the transaction is executed.
When reading a change-slot-values transaction in an anonymous
transaction that references an object that has been created in the
same transaction, it fails to find this new object and deserializes it
as NIL.

The solution to the problem is to actually execute the subtransactions
of an anonymous transaction while reading the transaction log instead
of first reading, then executing.  The downside of this is that if the
anonymous transaction was written incompletely, it will be partially
restored before the error is detected.  There needs to be a restart in
place for anonymous transactions that makes it possible to easily
truncate the txn log before the partially written anonymous
transaction begins so that the user has a way to restore to just that
point.