Added children-to-initforms to xml-import.xml and dtd-name to the xml-class
Assorted changes, mostly minimal. Got the initial import example in
the tutorial working.

Reorganised the xml-export facility, that now works for the examples in the tutorial

Created write-to-xml macro to wrap buil-xml (previously named write-to-xml) in order to maintain the API of the tutorial. Code now uses cxml:with-xml-output. Refactored the (now) build-xml methods: the test that dispatches to standard-objects or objects that are of the xml-class metaclass must be done inside the method.

Modified some minor details in the tutorial.

Changed the path of the imported DTDs to make it easier for people who are evaluating the code. Added :containment class option to make one of the examples work (it should be fixed in the code though). Substituted calls to * (SBCL was not cooperating with that).

Changed indentation tabs to spaces