Improvements in DTD checking and XML output
Added write-to-xml-string convenience macro, that returns a string
instead of using *standard-output*

Reintroduced two methods from previous bknr versions: elmdef-children
and get-dtd-elmdef, plus code to xml-finalize. This part of the code
was responsible for some sanity checks related to the slots/element
mapping and validation of containments.

Added some documentation strings to the reintroduced methods.