Rebuild indexes when finalize-inheritance is called
In particular, on Lispworks, if you have a class A, and a subclass B,
if you reinitialize A, then reinitialize-instance is not called on
B. However, compute-slots and finalize-inheritance *are* called on B.

The call to compute-slots creates new index objects in
indexed-class-indices. But these index-objects don't get reinitialized
in the previous version. In the current version it does get
reinitialized.

There's still a minor bug: We might create multiple index-objects for
the slot for B, but AFAICT, the final behavior is still correct even
if it doesn't have optimal performance. The redundant indices will
resolve itself the next time the class B is directly reinitialized.