Accommodate for SBCL feature absence on Win32. N.B. Users on Win.
The sb-posix interface.lisp in SBCL does not export the functions FSYNC or
TRUNCATE on the Microsoft Windows platform. The library fails to load on
Windows.

See:
https://github.com/sbcl/sbcl/blob/master/contrib/sb-posix/interface.lisp#L179

As the library author explains,

The call to fsync is optional in the sense that it won't make any functional
difference if you don't call it. The idea behind it is that after fsync has been
invoked, all file updates have been written to stable storage, which is not
quite true even on Linux anyway.

The call to truncate is required when restoring from the log to a certain point
in time or when there is an error during restore. It is not needed during normal
operation.

NOTE FOR those on Win: You'll lose the ability to do point-in-time restores and
automatic restores will become less robust.